requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Babel==2.12.1
backports.zoneinfo==0.2.1
Django==4.1.7
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
phonenumbers==8.13.8
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
